# code/generated/test_service_server/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: avelasco.horcajada@knowmadmood.com
    name: Antonio Velasco Horcajada
  description: "API for authoring immutable test definitions, composing test plans,\
    \ executing validated plans, and publishing read-only projections to an external\
    \ testing viewer."
  title: Test Service
  version: 0.1.1
servers:
- description: Local development server
  url: http://localhost:8080
security:
- bearerAuth: []
...
paths:
  /v1/projects:
    get:
      description: Returns a paginated list of active projects registered in the local
        Project Catalog.
      operationId: listProjects
      parameters:
      - description: Number of records to skip before returning results.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          $ref: "#/components/schemas/Offset"
        style: form
      - description: Maximum number of records returned in one page.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          $ref: "#/components/schemas/Limit"
        style: form
      - description: Field used to sort the result set.
        explode: true
        in: query
        name: sortBy
        required: false
        schema:
          default: version
          enum:
          - status
          - version
          - name
          example: version
          type: string
        style: form
      - description: Sort direction.
        explode: true
        in: query
        name: order
        required: false
        schema:
          $ref: "#/components/schemas/SortOrder"
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
          description: Active projects.
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: List projects
      tags:
      - Projects
    post:
      description: Registers a Jira project after a read-only validation that the
        exact key exists and is active in Jira.
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
          description: Project registered.
          headers:
            Location:
              description: URI reference of the created resource.
              explode: false
              required: true
              schema:
                format: uri-reference
                type: string
              style: simple
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: Register a project
      tags:
      - Projects
  /v1/projects/{projectKey}:
    delete:
      description: "Marks the project as DELETED without deleting tests, sets, plans,\
        \ executions, results, or artifacts."
      operationId: deleteProject
      parameters:
      - description: Stable key of the project in the Project Catalog.
        explode: false
        in: path
        name: projectKey
        required: true
        schema:
          $ref: "#/components/schemas/ProjectKey"
        style: simple
      responses:
        "204":
          description: Project logically deleted. No response body is returned.
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: Logically delete a project
      tags:
      - Projects
    get:
      description: "Returns a project by its stable project key, including logically\
        \ deleted projects when explicitly authorized."
      operationId: getProject
      parameters:
      - description: Stable key of the project in the Project Catalog.
        explode: false
        in: path
        name: projectKey
        required: true
        schema:
          $ref: "#/components/schemas/ProjectKey"
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
          description: Project.
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: Get a project
      tags:
      - Projects
...
components:
...
  schemas:
    Offset:
      default: 0
      description: Number of records to skip before returning results.
      format: int32
      minimum: 0
      title: Offset
      type: integer
    Limit:
      default: 20
      description: Maximum number of records returned in one page.
      format: int32
      maximum: 100
      minimum: 1
      title: Limit
      type: integer
    SortOrder:
      default: ASC
      description: Sort direction for ordered collection results.
      enum:
      - ASC
      - DESC
      example: ASC
      title: SortOrder
      type: string
    ProjectKey:
      description: Stable key used to identify a project in the Project Catalog.
      example: IAG
      maxLength: 20
      minLength: 2
      title: ProjectKey
      type: string
    Project:
      additionalProperties: false
      description: Project Catalog entry used to scope all test resources.
      example:
        key: IAG
        name: AI Gateway
        status: ACTIVE
        createdAt: 2026-01-15T10:30:00Z
        createdBy: admin@example.com
        deletedAt: 2026-09-18T10:30:00Z
        deletedBy: admin@example.com
      properties:
        key:
          description: Stable key used to identify a project in the Project Catalog.
          example: IAG
          maxLength: 20
          minLength: 2
          title: ProjectKey
          type: string
        name:
          description: Human-readable project name managed by Test Service.
          example: AI Gateway
          maxLength: 255
          minLength: 1
          title: name
          type: string
        status:
          description: Project lifecycle status. DELETED is a logical deletion state.
          enum:
          - ACTIVE
          - DELETED
          example: ACTIVE
          title: status
          type: string
        createdAt:
          description: UTC timestamp when the project was created.
          example: 2026-01-15T10:30:00Z
          format: date-time
          title: createdAt
          type: string
        createdBy:
          description: Authenticated administrator that created the project.
          example: admin@example.com
          maxLength: 255
          minLength: 1
          title: createdBy
          type: string
        deletedAt:
          description: UTC timestamp when the project was logically deleted.
          example: 2026-09-18T10:30:00Z
          format: date-time
          title: deletedAt
          type: string
        deletedBy:
          description: Authenticated administrator that logically deleted the project.
          example: admin@example.com
          maxLength: 255
          minLength: 1
          title: deletedBy
          type: string
      required:
      - createdAt
      - createdBy
      - key
      - name
      - status
      title: Project
      type: object
    Pagination:
      additionalProperties: false
      description: Offset pagination metadata for a collection response.
      example:
        offset: 0
        limit: 60
        total: 125
      properties:
        offset:
          default: 0
          description: Number of records to skip before returning results.
          format: int32
          minimum: 0
          title: Offset
          type: integer
        limit:
          default: 20
          description: Maximum number of records returned in one page.
          format: int32
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
        total:
          description: Total number of records matching the query.
          example: 125
          format: int64
          minimum: 0
          title: total
          type: integer
      required:
      - limit
      - offset
      - total
      title: Pagination
      type: object
    ProjectListResponse:
      additionalProperties: false
      description: Paginated collection of active projects in the local Project Catalog.
      example:
        data: []
        pagination:
          offset: 0
          limit: 20
          total: 0
      properties:
        data:
          description: Active projects in the requested page.
          items:
            $ref: "#/components/schemas/Project"
          maxItems: 100
          type: array
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
      - data
      - pagination
      title: ProjectListResponse
      type: object
    ErrorDetails:
      properties:
        type:
          default: about:blank
          description: |
            URI reference identifying the problem type. When omitted, "about:blank" is assumed.
          example: /problems/invalid-request
          format: uri-reference
          title: type
          type: string
        status:
          description: |
            HTTP status code generated by the server for this occurrence of the problem.
          example: 400
          format: int32
          maximum: 599
          minimum: 100
          title: status
          type: integer
        title:
          description: |
            Short, human-readable summary of the problem type. It should remain stable across occurrences of the same problem.
          example: Bad Request
          maxLength: 1000
          title: title
          type: string
        detail:
          description: |
            Human-readable explanation specific to this occurrence of the problem.
          example: The request is invalid.
          maxLength: 5000
          title: detail
          type: string
        instance:
          description: |
            URI reference identifying this specific occurrence of the problem. It may be used to correlate the response with logs or traces.
          example: /errors/01K5A8F7Q
          format: uri-reference
          title: instance
          type: string
      required:
      - detail
      - status
      - title
      title: ErrorDetails
      type: object
    CreateProjectRequest:
      additionalProperties: false
      description: Administrative request to register a Jira project in the local
        Project Catalog.
      example:
        key: IAG
        name: AI Gateway
      properties:
        key:
          description: Stable key used to identify a project in the Project Catalog.
          example: IAG
          maxLength: 20
          minLength: 2
          title: ProjectKey
          type: string
        name:
          description: Local human-readable name for the project.
          example: AI Gateway
          maxLength: 255
          minLength: 1
          title: name
          type: string
      required:
      - key
      - name
      title: CreateProjectRequest
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Authentication using a bearer token.
      scheme: bearer
      type: http