# code/generated/test_service_server/test_service_server/models/sort_order.py
# coding: utf-8

"""
    Test Service

    API for authoring immutable test definitions, composing test plans, executing validated plans, and publishing read-only projections to an external testing viewer.

    The version of the OpenAPI document: 0.1.1
    Contact: avelasco.horcajada@knowmadmood.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class SortOrder(str, Enum):
    """
    Sort direction for ordered collection results.
    """

    """
    allowed enum values
    """
    ASC = 'ASC'
    DESC = 'DESC'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of SortOrder from a JSON string"""
        return cls(json.loads(json_str))